refactor: Update Dockerfile and package.json to use new production startup command
The Dockerfile and package.json have been modified to use the new production startup command `npm run start:migrate:prod`. This command includes the Prisma migration deployment before starting the application in production mode. This change ensures that the application is always up-to-date with the latest database schema changes when running in production.