refactor: Add healthcheck configuration to docker-compose.yml
The docker-compose.yml file has been updated to include a healthcheck configuration for the service. This configuration runs a healthcheck command every 10 seconds to check if the PostgreSQL database is ready to accept connections. If the healthcheck fails, the service will be restarted after 30 seconds. This change improves the reliability of the application by ensuring that the database is available before starting the service.