refactor: Include thumbnail and published fields in createArticle method
This commit updates the `ArticleService` to include the `thumbnail` and `published` fields in the `createArticle` method. If the `payload` object contains a `thumbnail` property, it will be included in the created article. Similarly, if the `payload` object contains a `published` property, it will be included in the created article. This change enhances the functionality of the `ArticleService` by allowing the creation of articles with thumbnail images and specifying their published status.