refactor: Update mydata method in UserController to include user authentication
This commit updates the `mydata` method in the `UserController` to include user authentication. It adds the `@UseGuards(AuthGuard)` decorator to the method and modifies the method signature to accept the `@Request()` object with the `user` property. This change improves the security of the application by requiring authentication for accessing the user's data and articles.