refactor: Add signOut method to AuthService for user logout
This commit adds a new method `signOut` to the `AuthService` for user logout functionality. The method takes in the `@Res()` object and the `token` as parameters. It adds the provided token to the token blacklist, clears the access token and user name cookies from the response, and returns a JSON response with a success message. This change enhances the security of the application by allowing users to securely log out and invalidating their access token.