refactor: Update readMyArticles method to return IResponse<ArticleWithAuthor[]> and include author information
This commit updates the `readMyArticles` method in the `ArticleController` and `ArticleService` to return an `IResponse<ArticleWithAuthor[]>` instead of directly returning an array of articles. The `ArticleWithAuthor` type is added to the `types/types.ts` file to include the author information along with the article data. The `getMyArticles` method in the `ArticleService` now fetches the articles and their corresponding authors, enhancing the functionality of the application by providing users with a way to access their own articles along with the author details.